Getters, setters, constructors, etc

+++ src/main/java/com/miguelprojects/travel_agency/Repository/TravelRepository.java
@@ -3,5 +3,5 @@ package com.miguelprojects.travel_agency.Repository;
 import com.miguelprojects.travel_agency.Models.Travel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TravelRepository extends JpaRepository<Travel, Integer> {
+public interface TravelRepository extends JpaRepository<Travel, Long> {
 }
